% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nprobust.R
\name{exclusion_mods}
\alias{exclusion_mods}
\title{Exclude all permutations of variables}
\usage{
exclusion_mods(vec, always_include = NULL, baseline_model, ...)
}
\arguments{
\item{vec}{A vector of strings giving variable names to be excluded from the model.}

\item{always_include}{A vector of strings giving variable names that should always be included in the model.  Defaults to `NULL`.}

\item{baseline_model}{A model object that will be updated with new formula.  As such, it must be compatible with `update()`.}

\item{...}{Other arguments, currently not used.}
}
\description{
Update model by excluding all permutations and combinations of indicated variables.
}
\examples{
data(mtcars)
base_mod <- lm(qsec ~ wt + hp + disp, data = mtcars)
# build models
mods <- exclusion_mods(c("cyl", "vs", "carb"),
                       always_include = c("wt", "hp", "disp"),
                       base_mod)
np_robust(base_mod, mods, vbl=c("wt", "hp", "disp"), type="fd")
}
