% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nprobust.R
\name{get_anova}
\alias{get_anova}
\title{Get F-test Results}
\usage{
get_anova(x, vbl, ...)
}
\arguments{
\item{x}{A model object}

\item{vbl}{A string giving the variable name associated with the effect of interest}

\item{...}{other arguments, not currently implemented.}
}
\value{
A scalar giving the p-value for the variable of interest.
}
\description{
Get the results of an F-test for dropping each model term independently.
}
\examples{
data(mtcars)
mod <- lm(mpg ~ poly(wt, 2) + hp + vs, data=mtcars)
get_anova(mod, "wt")
}
