% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nprobust.R
\name{make_ME_mods}
\alias{make_ME_mods}
\title{Makes models where the confounding variable's correlation with the 
variable of interest is controlled.}
\usage{
make_ME_mods(
  model,
  data,
  vbl,
  scale_dv = TRUE,
  scale_vbl = TRUE,
  n_levels = 21,
  ...
)
}
\arguments{
\item{model}{A model object of class - `lm` and `glm` should work, though `glm` will 
likely require `scale_dv=FALSE`.}

\item{data}{A data frame from which model variables can be retrieved.}

\item{vbl}{A string giving the name of the variable to be evaluated.}

\item{scale_dv}{If `scale_dv=TRUE` and `scale_vbl=FALSE`, then the dependent variable
in the model will be scaled to have the same variance as the variable of interest. If
`scale_dv=TRUE` and `scale_vbl=TRUE`, then both variables will be scaled to have variance of 1.}

\item{scale_vbl}{If `scale_dv=FALSE` and `scale_vbl=TRUE`, then the variable of interest
in the model will be scaled to have the same variance as the dependent variable. If
`scale_dv=TRUE` and `scale_vbl=TRUE`, then both variables will be scaled to have variance of 1.}

\item{n_levels}{The number of levels used in the weights between the common variance and error.}

\item{...}{other arguments, not currently implemented.}
}
\value{
A list with three elements: 
* params: a data frame with the variables returned by `broom::tidy()` along with `rzx` and `rzy` #' - the correlation between the confounding variable and the independent and dependent variables as #' well as `p_anova` the p-value for the F-test of the joint significance of the regressors 
belonging to the variable of interest. 
* models: a list of the models estimated. 
* data: A data frame containing the variables for the models.
}
\description{
Makes models where the confounding variable's correlation with the 
variable of interest is controlled.
}
\examples{
data(mtcars)
base_mod <- lm(mpg ~ disp + wt + hp + vs, data=mtcars)
m_hp <- make_ME_mods(base_mod, mtcars, "hp", scale_dv=TRUE, scale_vbl=TRUE)
}
