% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nprobust.R
\name{np_robust}
\alias{np_robust}
\title{Neumayer-Plumper Robustness}
\usage{
np_robust(
  base_mod,
  robust_mod,
  vbl,
  base_args = list(),
  robust_args = list(),
  type = c("fd", "slope", "pred"),
  ...
)
}
\arguments{
\item{base_mod}{A baseline model object}

\item{robust_mod}{An alternative model object or a list of model objects against which robustness of the baseline model will be evaluated.}

\item{vbl}{A character string giving the variable
whose robustness is being tested.}

\item{base_args}{Arguments to be passed to `avg_slopes()` or `avg_comparisons()` that govern the baseline model effects.}

\item{robust_args}{Arguments to be passed to `avg_slopes()` or `avg_comparisons()` that govern the baseline model effects; these should produce the same number of
effect estimates as the base_args specification does.}

\item{type}{The quantity being tested - first difference "fd",  marginal effect "slope", or prediction "pred".}

\item{...}{Other arguments to be passed down, not implemented.}
}
\value{
A data frame that contains the robust model estimate and standard error along with the 95% CI from the baseline model and the robustness calculation.
}
\description{
Calculate the Neumayer and Pluemper (2017) measure
of robustness for a robustness model relative to
a baseline model
}
