% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nprobust.R
\name{rob_gamlss}
\alias{rob_gamlss}
\title{Estimate GAMLSS with Robustness Weights}
\usage{
rob_gamlss(formula, data, ...)
}
\arguments{
\item{formula}{The formula for the mean in a gamlss model}

\item{data}{A data frame that can be used to fit the model.  As it is being used in a gamlss model, it cannot contain any missing data.}

\item{...}{Other arguments that will be passed down to `gamlss()`.}
}
\value{
A list with elements `mod` which holds the `gamlss` model object and `weights` which give the robust regression weights
}
\description{
Estimates the GAMLSS model with an iterative step that
identifies outliers and downweights them according
to Huber's Bisquare function.
}
\details{
The function creates a variable called `weight` in the dataset that holds the robust regression weights.  If there is an existing
variable in your data called `weight`, it will be overwritten.
}
\examples{
\dontrun{
data(mtcars)
baseg <- gamlss::gamlss(qsec ~ wt + hp + disp + vs + carb, data = mtcars)
robg <- rob_gamlss(qsec ~ wt + hp + disp + vs + carb, data = mtcars)
summary(robg$mod)
summary(robg$weights)
}
}
