% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nprobust.R
\name{sim_data}
\alias{sim_data}
\alias{sim_data.lm}
\alias{sim_data.glm}
\title{Simulate data under the base model specification}
\usage{
sim_data(model, orig_data, R = 100, ...)

\method{sim_data}{lm}(model, orig_data, R = 100, ...)

\method{sim_data}{glm}(model, orig_data, R = 100, ...)
}
\arguments{
\item{model}{The baseline model for a robustness analysis.}

\item{orig_data}{The data frame used to fit the model.}

\item{R}{Number of simulated datasets to make.}

\item{...}{Other arguments, currently not implemented.}
}
\description{
Simulates data for lm or glm (binomial or poisson families) assuming that the
baseline model is correctly specified.
}
\examples{
data(mtcars)
mod <- lm(qsec ~ wt + disp + hp, data=mtcars)
sdat <- sim_data(mod, mtcars, R=1)
head(sdat[[1]])
}
