% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nprobust.R
\name{sim_models}
\alias{sim_models}
\alias{sim_models.lm}
\alias{sim_models.glm}
\title{Simulate model results}
\usage{
sim_models(model, R = 100, ...)

\method{sim_models}{lm}(model, R = 100, ...)

\method{sim_models}{glm}(model, R = 100, ...)
}
\arguments{
\item{model}{A model object of class `lm` or `glm`.}

\item{R}{Integer giving the number of times the model should be simulated.}

\item{...}{Other arguments, currently not used.}
}
\description{
Simulates model results under the assumption that the input model 
is correct.
}
\examples{
data(mtcars)
lmod <- lm(qsec ~ hp + wt + vs, data=mtcars)
mods <- sim_models(lmod, R=10)
}
