% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nprobust.R
\name{sim_robust}
\alias{sim_robust}
\title{Simulate robustness scores under baseline model}
\usage{
sim_robust(
  base_model,
  robust_models,
  orig_data,
  R = 100,
  rob_type = c("avg", "ind"),
  vbl = NULL,
  type = c("fd", "slope", "pred"),
  base_args = NULL,
  robust_args = NULL,
  show_progress = TRUE,
  arrange_robust = TRUE,
  ...
)
}
\arguments{
\item{base_model}{A model, currently lm and glm (binomial and poisson families) are supported.}

\item{robust_models}{A model or list of models against which robustness will be evaluated}

\item{orig_data}{The data frame from which the data for `base_model` and `robust_models` are obtained.}

\item{R}{Number of simulations}

\item{rob_type}{Type of robustness analysis to do `"avg"` uses the `np_robust()` function and `"ind"` uses the `ind_robust()` function.}

\item{vbl}{Variable(s) for which robustness is to be evaluated.}

\item{type}{Quantity for which robustness will be evaluated - first difference (`"fd"`), slope (`"slope"`) or predictions (`"pred"`).}

\item{base_args}{Arguments to be passed to the robustness analysis for the baseline model.}

\item{robust_args}{Arguments to be passed to the robustness analysis for the robustness models.}

\item{show_progress}{If `TRUE`, `pbapply::pblapply()` will be used instead of `lapply()`.}

\item{arrange_robust}{If `TRUE`, the data will be re-arranged according to average simulated robustness value and a new factor variable
`label` will be created that has the appropriate levels.}

\item{...}{Other arguments passed down to the robustness functions.}
}
\description{
Simulates robustness scores assuming that the baseline model is correctly
specified.
}
